requirements.txt
absl-py==0.14.1
asgiref==3.4.1
astunparse==1.6.3
bcrypt==3.2.0
cachetools==4.2.4
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.7
clang==5.0
click==8.0.1
colorgram.py==1.2.0
cycler==0.10.0
fastapi==0.68.1
flatbuffers==1.12
gast==0.4.0
google-auth==2.3.0
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
greenlet==1.1.2
grpcio==1.41.0
h11==0.12.0
h5py==3.1.0
httptools==0.2.0
idna==3.3
imageio==2.9.0
keras==2.6.0
Keras-Preprocessing==1.1.2
kiwisolver==1.3.2
Markdown==3.3.4
matplotlib==3.4.3
networkx==2.6.3
numpy==1.19.5
oauthlib==3.1.1
opencv-python-headless==4.5.3.56
opt-einsum==3.3.0
passlib==1.7.4
Pillow==8.3.2
protobuf==3.18.1
psycopg2==2.9.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
pydantic==1.8.2
pyparsing==2.4.7
python-dateutil==2.8.2
python-dotenv==0.19.0
python-multipart==0.0.5
PyWavelets==1.1.1
PyYAML==5.4.1
requests==2.26.0
requests-oauthlib==1.3.0
rsa==4.7.2
scikit-image==0.18.3
scipy==1.7.1
six==1.15.0
SQLAlchemy==1.4.25
starlette==0.14.2
tensorboard==2.7.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.0
tensorflow==2.6.0
tensorflow-estimator==2.6.0
tensorflow-hub==0.12.0
termcolor==1.1.0
tifffile==2021.10.12
torch==1.9.1
torchvision==0.10.1
typing-extensions==3.7.4.3
urllib3==1.26.7
uvicorn==0.15.0
uvloop==0.16.0
watchgod==0.7
websockets==10.0
Werkzeug==2.0.2
wrapt==1.12.1
